fix: moving to seperate components
