+++ GitHubIssueTracker/src/components/index.ts
@@ -2,3 +2,5 @@ export * from './buttons/DefaultButton';
 export * from './texts/HeaderText';
 export * from './IssuesComponents/listItem';
 export * from './Icons';
+export * from './IssuesComponents/issuesFilterButton';
+export * from './IssuesComponents/issueListView';
